requirements.txt
pandas
streamlit==1.34.0
#PyYAML==6.0.1
pyyaml==6.0.1
scikit-learn
#joblib==1.4.2
python-box==7.1.1
ensure==1.0.4
nltk==3.8.1
-e .